Consolidate Tag Request interfaces
