+++ ui/src/Tags/TagsModalContent.tsx
@@ -18,7 +18,7 @@
 import {JSX} from '@babel/types';
 import React, {useState} from 'react';
 import ConfirmationModal, {ConfirmationModalProps} from 'Modal/ConfirmationModal/ConfirmationModal';
-import {TagRequest} from './TagRequest.interface';
+import {TagRequest} from '../Types/TagRequest';
 import {createDataTestId} from '../Utils/ReactUtils';
 import ViewTagRow from '../ModalFormComponents/ViewTagRow';
 import EditTagRow from '../ModalFormComponents/EditTagRow';
